👽️ Implement modified EntityPositionS2CPacket format

+++ src/main/java/io/github/theepicblock/polymc/impl/poly/entity/MissingEntityPoly.java
@@ -32,7 +32,7 @@ public class MissingEntityPoly<T extends Entity> implements EntityPoly<T> {
 
         @Override
         public void onMove(PacketConsumer players) {
-            item.move(players, this.getPosition(), (byte)0, (byte)0, true);
+            item.move(players, this.getPosition(), 0, 0, true);
         }
 
         @Override
